requirements.txt
deepface
numpy
Pillow
PySimpleGUI
tensorflow
tf-keras